+++ CHANGELOG.md
@@ -6,9 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### 🚀 Features
+- Add update relations PUT to product associations resource ([#199](https://github.com/whirli/whirli-client/pull/199))
 ## [0.0.22] - 2020-06-09
+
 ### 🚀 Features
 - Add all product-lists and product-list-lines endpoints, both public and private ([#196](https://github.com/whirli/whirli-client/pull/196))
+
 ### 🐛 Bug fixes
 - Temp change returnOrder complete from POST to PUT to test staging ([#195](https://github.com/whirli/whirli-client/pull/195))