+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### 🚀 Features
-- Add preview change subscription endpoint
+- Add preview change subscription endpoint ([#201](https://github.com/whirli/whirli-client/pull/201))
 
 ### 🚀 Features
 - Add update relations PUT to product associations resource ([#199](https://github.com/whirli/whirli-client/pull/199))