+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### 🚀 Features
+- Add preview change subscription endpoint ([#201](https://github.com/whirli/whirli-client/pull/201))
 
 ### 🚀 Features
 - Add update relations PUT to product associations resource ([#199](https://github.com/whirli/whirli-client/pull/199))